app/requirements.txt
Django==2.2.7
django-cors-headers==3.2.0
django-debug-toolbar==2.1
django-filter==2.2.0
djangorestframework==3.10.3
Jinja2==2.10.3
python-dateutil==2.8.1
slackclient
psycopg2
requests
